Select the git repository more wisely
If the workspace has more than one directory, warn the users that the
extension could have picked the wrong one and work in unexpected ways.

If the workspace has only one directory, it can still contain multiple
git repositories. In that case, we explicitly try to retrieve the top
level git repository because it's the one in which we'll run git
format-patch commands anyway.

This should make the extension work in less mysterious ways for users
with multiple git repositories loaded at the same time.