Remember sent emails
This instruments the "git send-email" terminal using the "script"
command, writing into a named pipe so that the extension can snoop on
that terminal's output.

It detects sent Message-Ids and remembers them in the series such that
links to archives of discussions over previous iterations of this series
can be shown on the left panel.