Make the patchwork.open command detection more reliable
The previous commit tried to detect whether "patchwork.open" is
available but that logic only works after the extension has been
activated. (if the user clicked on the left icon of the patchwork
extension)

Instead, we can try to directly call the command, which might in turn
activate the extension and make it available. The result of this being a
"Thenable" means we can catch errors and fallback to opening an archive
link in this situation.